requirements.txt
django
django-mptt
pandas
psycopg2
tqdm